Test `LiveDataMapperTests` in isolation
