+++ HackrNews/Live Hackr News API/LiveDataMapper.swift
@@ -5,7 +5,7 @@
 import Foundation
 
 public enum LiveDataMapper {
-    public typealias LiveItem = Int
+    typealias LiveItem = Int
 
     public static func map(data: Data, response: HTTPURLResponse) throws -> [LiveHackrNew] {
         guard response.isOK, let data = try? JSONDecoder().decode([LiveItem].self, from: data) else {
